+++ packages/expo-modules-core/android/src/main/cpp/JSIContext.cpp
@@ -359,6 +359,7 @@ void JSIContext::prepareForDeallocation() {
   runtimeHolder.reset();
   jniDeallocator.reset();
   javaPart_.reset();
+  wasDeallocated = true;
 }
 
 void JSIContext::jniSetNativeStateForSharedObject(
@@ -384,6 +385,10 @@ void JSIContext::jniSetNativeStateForSharedObject(
     ->cthis()
     ->get()
     ->setNativeState(runtimeHolder->get(), std::move(nativeState));
+}
+
+bool JSIContext::isDeallocated() const {
+  return wasDeallocated;
 }
 
 thread_local std::unordered_map<uintptr_t, JSIContext *> jsiContexts;